build: move localStorage functions to /lib
